Grammar tweaks - Update README.md
Thank you for a great article based on an interesting idea.

I hit some speed bumps while reading this, so I took a few moments to help make the English more consistent and correct. Mostly it's typos with a few things like making the items in a list consistently starting with verbs, for example.

Feel free to take or reject any part of this, of course.

Enjoy!